Add sparsezoo utils stubs
